Move timezone and utils to separate package
