+++ tz/tz.go
@@ -1,0 +1,53 @@
+package tz
+
+import (
+	"time"
+
+	"github.com/raintank/dur"
+)
+
+var (
+	TimeZone *time.Location
+)
+
+type FromTo struct {
+	From  string `json:"from" form:"from"`
+	Until string `json:"until" form:"until"`
+	To    string `json:"to" form:"to"` // graphite uses 'until' but we allow to alternatively cause it's shorter
+	Tz    string `json:"tz" form:"tz"`
+}
+
+func GetFromTo(ft FromTo, now time.Time, defaultFrom, defaultTo uint32) (uint32, uint32, error) {
+	loc, err := getLocation(ft.Tz)
+	if err != nil {
+		return 0, 0, err
+	}
+
+	from := ft.From
+	to := ft.To
+	if to == "" {
+		to = ft.Until
+	}
+
+	fromUnix, err := dur.ParseDateTime(from, loc, now, defaultFrom)
+	if err != nil {
+		return 0, 0, err
+	}
+
+	toUnix, err := dur.ParseDateTime(to, loc, now, defaultTo)
+	if err != nil {
+		return 0, 0, err
+	}
+
+	return fromUnix, toUnix, nil
+}
+
+func getLocation(desc string) (*time.Location, error) {
+	switch desc {
+	case "":
+		return TimeZone, nil
+	case "local":
+		return time.Local, nil
+	}
+	return time.LoadLocation(desc)
+}
